PLF-7986 : set css-priority of customStyle to 0 to allow override by others custom module (#163)
The customStyle module does not have a css priority, meaning it is loaded with the highest priority (default behavior of skin service).
It means that others custom style modules cannot have a higher priority and will never be loaded after the customStyle css, which makes it harder to override.
The fix sets the customStyle priority to 0. Therefore, by default others custom module will have a higher priority.